全国地图读本地配置

+++ model/national_map.go
@@ -4,7 +4,7 @@ type NationalMap struct {
 	Id		int			`json:"id" xorm:"id pk autoincr"`
 	X		int			`json:"x"`
 	Y		int			`json:"y"`
-	Type	int8		`json:"type"`
+	Type	int			`json:"type"`
 	Level	int			`json:"level"`
 }
